Added ability to change map view with brounds
I would like to show multiple push pins on the map, and set the map view bounds to fit the pins. I can obtain the bounds that fits the pins easily with Maps API, however, there was no way to set the view to this bounds.

I don't want to convert the bounds to center and zoom level, so I created this patch for allowing me to set the map view to the bounds

I also updated the push pins exampel to demostrate how to fit the bounds of the map view to  multiple pins.